reference/odie-oaa/requirements.txt
# OAA Connector Dependencies
oaaclient>=1.0.0
python-dotenv>=1.0.0
requests>=2.28.0
